Time: 1 ms (99.97%), Space: 42.8 MB (26.71%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.java
@@ -1,6 +1,6 @@
 class Solution {
     public int reverse(int x) {
-        long rem=0;
+        int rem=0;
         long res=0;
         while (x!=0){
             rem=x%10;
